Removed extra water.swift file

+++ WeatherApp/Weather.swift
@@ -1,5 +1,5 @@
 //
-//  Forecast.swift
+//  Weather.swift
 //  WeatherApp
 //
 //  Created by Griffin Hammer on 12/2/15.
@@ -8,10 +8,11 @@
 
 import Foundation
 
+
 struct Weather {
     var icon : String?
-    var highTemp : Int?
-    var lowTemp : Int?
+    var highTemp : Float?
+    var lowTemp : Float?
     var desc : String?
     var day : String?
 }
